Move confirmation page to its controller
There will be multiple types of confirmation pages with slightly
different actions.
