+++ config/routes.rb
@@ -7,7 +7,9 @@ Rails.application.routes.draw do
 
   resource :summary, only: :show
 
-  resource :submission, only: [:create, :show]
+  resource :submission, only: :create
+
+  resource :confirmation, only: :show
 
   resource :session, only: :destroy do
     get :start
