+++ spec/controllers/submissions_controller_spec.rb
@@ -30,8 +30,8 @@ RSpec.describe SubmissionsController, type: :controller do
         expect(storage).to have_received(:submission_result=).with(response)
       end
 
-      it 'redirects to the show action' do
-        expect(response).to redirect_to(submission_path)
+      it 'redirects to the confirmation page' do
+        expect(response).to redirect_to(confirmation_path)
       end
     end
 
@@ -49,33 +49,6 @@ RSpec.describe SubmissionsController, type: :controller do
       it 'sets an flash error message' do
         expect(flash[:error]).to eql('We couldn\'t process your application at this time. Please try again later.')
       end
-    end
-  end
-
-  describe 'GET #show' do
-    let(:result) { { result: true, message: 'HWF-010101' } }
-    let(:storage) { double(submission_result: result, time_taken: 600, clear: nil) }
-
-    before do
-      allow(online_application).to receive(:benefits).and_return(true)
-
-      get :show
-    end
-
-    it 'renders the show template' do
-      expect(response).to render_template(:show)
-    end
-
-    it 'assigns the response object from the session' do
-      expect(assigns(:result)).to eql(result)
-    end
-
-    it 'clears the storage' do
-      expect(storage).to have_received(:clear)
-    end
-
-    it 'assigns the online application model' do
-      expect(assigns(:online_application)).to eql(online_application)
     end
   end
 end
